+++ stories/stories.module.scss
@@ -1,0 +1,91 @@
+@import '../src/assets/sass/index.scss';
+
+section {
+    padding: 2em;
+    z-index: $z-initials;
+}
+
+.section {
+    &-grey {
+        background: $color-background;
+    }
+
+    &-purple {
+        color: $color-white;
+        background: $color-purple;
+    }
+
+    &-black {
+        color: $color-white;
+        background: $color-base;
+    }
+
+    &-form-min {
+        max-width: $google-spacing * 50;
+    }
+
+    &-relative {
+        position: relative;
+    }
+}
+
+.icons-grid {
+    display: flex;
+    flex-wrap: wrap;
+    margin-bottom: 2em;
+    text-align: center;
+
+    &__item {
+        display: flex;
+        flex-flow: row wrap;
+        align-items: flex-end;
+        justify-content: center;
+        width: $google-spacing * 16;
+        height: $google-spacing * 8;
+    }
+
+    &__letter {
+        margin: 1em 0 0;
+        width: 100%;
+        color: calc-color(tint, $color-base, 30);
+        text-align: left;
+        font-weight: bold;
+        text-transform: uppercase;
+    }
+
+    &__icon,
+    &__text {
+        flex: 1 0 100%;
+    }
+
+    &__icon {
+        margin-top: 1rem;
+    }
+
+    &__text {
+        font-size: $piko-size;
+        color: calc-color(fade, $color-base, 50);
+    }
+
+    &--large {
+        .icons-grid__item {
+            height: $google-spacing * 16;
+        }
+    }
+}
+
+.snackbars {
+    .kui-snackbar__container {
+        position: relative;
+        left: 0;
+        transform: translateX(0);
+    }
+}
+
+.stories-loader-small {
+    display: inline-flex;
+    justify-content: center;
+    align-items: center;
+    height: $google-spacing * 5;
+    width: $google-spacing * 5;
+}
